Move query streaming client code to its own file
This will be shared by the RemoteButler query_all_datasets implementation in an upcoming commit.
